Add files via upload
This is a January arXiv paper about the new results of Quantum Point Contact (QPC) experiment in quantum Hall regime. They claimed to realise the braiding of fractional charged anyons in IQH when the filling factor is 2. The main results are the detection of cross-correlation and Fano factor. The theoretical model about the bosonization and refermionisation of the edge electrons are constructed in Supplementary information. After the procedure, interacting fermions can be regarded as free edge magnetoplasmons.